# docs/02_api/openapi.yaml
openapi: 3.0.3
info:
  title: Aviation CAMO & MRO API
  version: 0.2.37
  description: 'Contract baseline for multi-tenant CAMO + MRO + Logistics platform.
    Tenant isolation is implemented via PostgreSQL schema-per-tenant.

    '
servers:
- url: http://localhost:8000
  description: Local dev
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'OIDC access token (Keycloak). Tenant context is resolved from
        token claims.

        '
...
  schemas:
...
    Aircraft:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        tail_number:
          type: string
        aircraft_type:
          type: string
        msn:
          type: string
        status:
          type: string
          enum:
          - ACTIVE
          - STORED
          - RETIRED
      required:
      - id
      - tenant_id
      - tail_number
      - aircraft_type
      - status
...
security:
- bearerAuth: []
paths:
...
  /v1/camo/aircraft:
    get:
      tags:
      - CAMO
      summary: List aircraft
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Aircraft'
    post:
      tags:
      - CAMO
      summary: Create aircraft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tail_number:
                  type: string
                aircraft_type:
                  type: string
                msn:
                  type: string
                status:
                  type: string
                  enum:
                  - ACTIVE
                  - STORED
                  - RETIRED
              required:
              - tail_number
              - aircraft_type
              - status
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Aircraft'
...
  /v1/aircraft:
    get:
      tags: [Aircraft]
      summary: List aircraft visible to current tenant (owned + accessible via MRO access)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Aircraft'
    post:
      tags: [Aircraft]
      summary: Create aircraft (owner = current tenant)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AircraftCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Aircraft'
  /v1/aircraft/{aircraft_id}:
    get:
      tags: [Aircraft]
      summary: Get aircraft (owner or MRO access)
      parameters:
      - name: aircraft_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Aircraft'
    put:
      tags: [Aircraft]
      summary: Update aircraft (owner full; MRO limited to status_tech/notes)
      parameters:
      - name: aircraft_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
              - $ref: '#/components/schemas/AircraftUpdateOwner'
              - $ref: '#/components/schemas/AircraftUpdateMro'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Aircraft'
    delete:
      tags: [Aircraft]
      summary: Delete aircraft (owner only)
      parameters:
      - name: aircraft_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: No Content
  /v1/aircraft/{aircraft_id}/mro-access:
    get:
      tags: [Aircraft]
      summary: List MRO access relations (owner only)
      parameters:
      - name: aircraft_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MroAccess'
    post:
      tags: [Aircraft]
      summary: Grant MRO access (owner only)
      parameters:
      - name: aircraft_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MroAccessCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MroAccess'
...
  /v1/aircraft/{aircraft_id}/maintenance-events:
    get:
      tags: [Aircraft]
      summary: List maintenance events (owner or assigned MRO)
      parameters:
      - name: aircraft_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AircraftMaintenanceEventOut'
    post:
      tags: [Aircraft]
      summary: Create maintenance event (owner only)
      parameters:
      - name: aircraft_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AircraftMaintenanceEventCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AircraftMaintenanceEventOut'
  /v1/aircraft/{aircraft_id}/maintenance-events/{event_id}:
    put:
      tags: [Aircraft]
      summary: Update maintenance event (owner or assigned MRO)
      parameters:
      - name: aircraft_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: event_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AircraftMaintenanceEventUpdate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AircraftMaintenanceEventOut'
    delete:
      tags: [Aircraft]
      summary: Delete maintenance event (owner or assigned MRO)
      parameters:
      - name: aircraft_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: event_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: No Content
...